Feat: Refine messaging flow
